requirements.txt
argh==0.26.2
asgiref==3.5.0
config2py==0.0.3
coverage==6.3.1
Django==4.0.2
django-extensions==3.1.5
django-jazzmin==2.4.9
django-mathfilters==1.0.0
django-sass-processor==1.1
djangorestframework==3.13.1
dol==0.1.71
gunicorn==20.1.0
importlib-resources==5.4.0
psycopg2-binary==2.9.3
py2store==0.1.11
pytz==2021.3
sqlparse==0.4.2

celery~=5.3.4